tutoriais_tarefas/requirements/local.txt
-r common.txt
ruff
pre-commit
pytest-django


djangorestframework-stubs
django-filter-stubs
django-stubs[compatible-mypy]
